% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{month}
\alias{month}
\title{month}
\usage{
month(month = 1)
}
\arguments{
\item{month}{Generate poem for which month. The default value is 1. The month range is from 1 to 12.}
}
\description{
\code{month} generates poem for each month
}
\examples{

month(1)

}
