% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreScreen.R
\name{Screen_X}
\alias{Screen_X}
\title{Screen features that are pure noise}
\usage{
Screen_X(X, max_prop = 0.5)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} data matrix.}

\item{max_prop}{A numeric value between [0, 1] specifying the maximal
proportional of pure noise features. Default is 0.5 meaning that at most
50\% of features are pure noise.}
}
\value{
A vector of indices that are detected as pure noise.
}
\description{
\code{Screen_X} finds features that are close to pure noise.
}
